adding the grid UI
